refactor(pg-proto): architect audit batch C — typed narrow ints + AuthSubCodeClass
Acting on architect findings #1, #4/#5, #66. Skipped #2 (payload_len
usize→u16) as triple-check showed the fallible-narrowing complexity
outweighs the 6-byte-per-variant saving on a size already budgeted.

# #1 — AuthSubCodeClass on UnsupportedAuthMethod.sub_code

Previously `sub_code: u32`. The u32 carried EITHER an unknown
wire code OR a known-but-wrong-for-state code (`.raw()` from
AuthSubCode at dispatch time). Both cases collapsed into one
untyped diagnostic.

Now:
```rust
pub enum AuthSubCodeClass {
    Unknown(u32),                              // server wire violation
    KnownButWrong(crate::wire::AuthSubCode),   // legal code, wrong state
}

UnsupportedAuthMethod { sub_code: AuthSubCodeClass }
```

Preserves the typed enum in the common case (server insisted on
SASL while client is in Trust auth — "Sasl wire value 10"
diagnostic instead of "unsupported auth method 10"). Tier-3
audit → tier-2 structural. No runtime cost: both variants fit in
8 bytes (discriminant + u32, aligned).

# #4/#5 — column_idx: usize → u8 in DecodeError + ColumnsIter

`DecodeError::TruncatedColumnLen`, `::NegativeColumnLength`,
`::TruncatedColumnData` all carried `column_idx: usize`. Invariant:
`column_idx < MAX_ROW_COLUMNS = 32` — fits `u8` with 8× headroom.

Narrowed to u8. Saves 7 bytes × 3 variants in the `DecodeError`
enum's size_of. `ColumnsIter.column_idx` field also narrowed.

Tier-3 audit → tier-2 structural (u8 caps the invariant at 255 ≫ 32).

# #66 — SessionParams.application_name capacity 64 → 128

Client-side `ApplicationName` is `MAX_APP_NAME_LEN = 128`. Server
echoes it back in `ParameterStatus`. But `SessionParams.application_name`
was `BoundedStr<64>` — a client sending a legitimate 100-byte
deployment-tagged name (`myapp-worker-pod-<sha256-truncated>`)
would see it truncated to 63+"…" on the session-params side —
fidelity loss on the round-trip.

Bumped to `BoundedStr<128>` for symmetric fidelity. No size-cap
regression (SessionParams size cap in lib.rs covers the 128-byte
form already).

Tests: 125 (unchanged — structural bump, existing integration
tests passed without adjustment). Clippy clean.

+++ crates/bsql-pg-proto/src/decode.rs
@@ -299,17 +299,18 @@ pub enum DecodeError {
     /// header. Malformed frame.
     TruncatedRow,
     /// A column's 4-byte length prefix is missing (fewer bytes
-    /// remain than expected). `column_idx` is 0-based.
+    /// remain than expected). `column_idx` is 0-based, bounded by
+    /// [`MAX_ROW_COLUMNS`] = 32 — fits `u8` with headroom.
     TruncatedColumnLen {
         /// Zero-based column index where the truncation was detected.
-        column_idx: usize,
+        column_idx: u8,
     },
     /// A column's declared length prefix is negative and is not the
     /// sentinel `-1` (which encodes SQL `NULL`). Other negative
     /// values are wire-level invalid.
     NegativeColumnLength {
         /// Zero-based column index.
-        column_idx: usize,
+        column_idx: u8,
         /// The offending length value.
         length: i32,
     },
@@ -317,7 +318,7 @@ pub enum DecodeError {
     /// prefix (partial row).
     TruncatedColumnData {
         /// Zero-based column index.
-        column_idx: usize,
+        column_idx: u8,
         /// Length declared by the prefix.
         declared_len: usize,
         /// Bytes actually remaining in the row body.
@@ -439,7 +440,7 @@ impl<'a> DataRowRef<'a> {
         ColumnsIter {
             remaining,
             columns_left: self.n_columns,
-            column_idx: 0,
+            column_idx: 0u8,
         }
     }
 }
@@ -461,7 +462,9 @@ impl<'a> DataRowRef<'a> {
 pub struct ColumnsIter<'a> {
     remaining: &'a [u8],
     columns_left: u16,
-    column_idx: usize,
+    /// Zero-based column index, bounded by [`MAX_ROW_COLUMNS`] = 32 —
+    /// `u8` with headroom. Propagated into `DecodeError::TruncatedColumn*`.
+    column_idx: u8,
 }
 
 impl<'a> Iterator for ColumnsIter<'a> {
